reactor.cc: Avoid using namespace internal
To make this file compile w/o imlicit namespace usage it takes explicit
internal:: scope in some places and few more.

In particular -- io_request::opname is moved to io_queue.cc and all the
code related to cpu stall detector is wrapped into namespace internal {}
block right in the reactor.cc file
